Reformatted tokenLayer
